Use improved ConstantEvaluator implementation
